Adds MetadataManager (#27)
* Adds MetadataManager

* feat: Refactor MetadataManager to use static methods for registration and retrieval

The code changes in `MetadataManager.php` remove the unnecessary comments and update the `register` and `get` methods to be static. This refactor improves the code readability and maintainability by clearly indicating that these methods can be accessed directly on the class without needing an instance.

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>

* feat: Update MetadataManager to use static methods for registration and retrieval

* feat: Update ClientTest to use service parameter in ClientFactory

The code changes in `ClientTest.php` update the `ClientFactory::create()` method to pass the `this->service` parameter as the first argument instead of `'test'`. This change ensures that the correct service is used when creating a new instance of the client.

---------

Co-authored-by: Deeka Wong <8337659+huangdijia@users.noreply.github.com>